+++ malleefowl/processes/wps_download.py
@@ -11,7 +11,7 @@ from pywps.app.Common import Metadata
 from malleefowl.download import download_files
 
 import logging
-LOGGER = logging.getLogger(__name__)
+LOGGER = logging.getLogger("PYWPS")
 
 
 class Download(Process):
@@ -25,7 +25,8 @@ class Download(Process):
                          max_occurs=1024,
                          ),
             ComplexInput('credentials', 'X509 Certificate',
-                         abstract='Optional X509 proxy certificate to access ESGF data.',
+                         abstract='Optional X509 proxy certificate to access ESGF data.'
+                         'This parameter is deprecated. Use X-X509-User-Proxy header variable.',
                          metadata=[Metadata('Info')],
                          min_occurs=0,
                          max_occurs=1,
@@ -60,11 +61,15 @@ class Download(Process):
         urls = [resource.data for resource in request.inputs['resource']]
         LOGGER.debug("downloading urls: %s", len(urls))
 
-        if 'credentials' in request.inputs:
+        if 'X-X509-User-Proxy' in request.http_request.headers:
+            credentials = request.http_request.headers['X-X509-User-Proxy']
+            LOGGER.debug('Using X509_USER_PROXY.')
+        elif 'credentials' in request.inputs:
             credentials = request.inputs['credentials'][0].file
-            LOGGER.debug('Using credentials.')
+            LOGGER.warn('Using deprecated input parameter credentials.')
         else:
             credentials = None
+            LOGGER.debug('Using no credentials')
 
         def monitor(msg, progress):
             response.update_status(msg, progress)